+++ lib/plausible_web/components/generic.ex
@@ -54,6 +54,17 @@ defmodule PlausibleWeb.Components.Generic do
     "md" => "btn-md"
   }
 
+  attr(:class, :string, default: "")
+  slot(:inner_block, required: true)
+
+  def auth_container(assigns) do
+    ~H"""
+    <div class={["w-full max-w-md mx-auto mt-10 pb-16 px-4", @class]}>
+      {render_slot(@inner_block)}
+    </div>
+    """
+  end
+
   attr(:type, :string, default: "button")
   attr(:theme, :string, default: "primary")
   attr(:size, :string, default: "md")