+++ test/plausible_web/controllers/admin_auth_controller_test.exs
@@ -17,7 +17,7 @@ defmodule PlausibleWeb.AdminAuthControllerTest do
 
       # "first launch" takes precedence
       conn = get(conn, "/register")
-      assert html_response(conn, 200) =~ "Enter your details"
+      assert html_response(conn, 200) =~ "Create your Plausible CE account"
     end
   end